Combined API call in coroutine. Coroutine dependencies added

+++ app/src/main/java/com/terentiev/coronatracker/data/Country.kt
@@ -3,13 +3,13 @@ package com.terentiev.coronatracker.data
 data class Country(
     val country: String,
     val countryInfo: CountryInfo,
-    val cases: Long,
-    val todayCases: Long,
-    val deaths: Long,
-    val todayDeaths: Long,
-    val recovered: Long,
-    val active: Long,
-    val critical: Long,
+    val cases: Int,
+    val todayCases: Int,
+    val deaths: Int,
+    val todayDeaths: Int,
+    val recovered: Int,
+    val active: Int,
+    val critical: Int,
     val casesPerOneMillion: Double,
     val deathsPerOneMillion: Double
 )
